会員登録時にデータベースのuserテーブルに会員情報が追加されない不具合の修正

+++ WEB-INF/src/servlet/MemberDeleteServlet.java
@@ -9,9 +9,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import dao.MemberDAO;
+import dao.UserDAO;
+import model.Member;
 
 @WebServlet("/memberDelete")
 public class MemberDeleteServlet extends HttpServlet {
+
     private static final long serialVersionUID = 1L;
 
     protected void doGet(
@@ -19,17 +22,45 @@ public class MemberDeleteServlet extends HttpServlet {
             HttpServletResponse response)
             throws ServletException, IOException {
 
-        int id = Integer.parseInt(request.getParameter("id"));
-
-        MemberDAO memberDAO = new MemberDAO();
-
         try {
-            memberDAO.delete(id);
+            int id =
+                    Integer.parseInt(
+                            request.getParameter("id"));
+
+            MemberDAO memberDAO =
+                    new MemberDAO();
+
+            /*
+             * 削除対象の会員を取得し、
+             * 社員番号を確認する。
+             */
+            Member member =
+                    memberDAO.findById(id);
+
+            if (member != null) {
+
+                UserDAO userDAO =
+                        new UserDAO();
+
+                /*
+                 * usersテーブルを先に削除する。
+                 */
+                userDAO.deleteByEmployeeNo(
+                        member.getEmployeeNo());
+
+                /*
+                 * users削除後にmembersを削除する。
+                 */
+                memberDAO.delete(id);
+            }
+
+            response.sendRedirect(
+                    request.getContextPath()
+                    + "/memberList");
+
         } catch (Exception e) {
+            e.printStackTrace();
             throw new ServletException(e);
         }
-
-        response.sendRedirect(
-                request.getContextPath() + "/memberList");
     }
 }
